refactor: ExchangeRoyalty
