+++ packages/marketplace/test/Exchange.test.ts
@@ -6,7 +6,6 @@ import {
   AssetERC721,
   AssetERC1155,
   FeeRecipientsData,
-  LibPartData,
   Asset,
 } from './utils/assets.ts';
 
@@ -31,11 +30,8 @@ describe('Exchange.sol', function () {
     ERC20Contract: Contract,
     ERC20Contract2: Contract,
     ERC721Contract: Contract,
-    ERC721WithRoyalty: Contract,
     ERC721WithRoyaltyV2981: Contract,
-    ERC721WithRoyaltyWithoutIROYALTYUGC: Contract,
     ERC1155Contract: Contract,
-    RoyaltiesProvider: Contract,
     protocolFeePrimary: number,
     protocolFeeSecondary: number,
     defaultFeeReceiver: Signer,
@@ -43,7 +39,6 @@ describe('Exchange.sol', function () {
     taker: Signer,
     admin: Signer,
     user: Signer,
-    deployer: Signer,
     makerAsset: Asset,
     takerAsset: Asset,
     orderLeft: Order,
@@ -64,19 +59,15 @@ describe('Exchange.sol', function () {
       ERC20Contract,
       ERC20Contract2,
       ERC721Contract,
-      ERC721WithRoyalty,
       ERC721WithRoyaltyV2981,
-      ERC721WithRoyaltyWithoutIROYALTYUGC,
       ERC1155Contract,
       protocolFeePrimary,
       protocolFeeSecondary,
       defaultFeeReceiver,
-      RoyaltiesProvider,
       user1: maker,
       user2: taker,
       admin,
       user,
-      deployer,
       EXCHANGE_ADMIN_ROLE,
       PAUSER_ROLE,
       ERC1776_OPERATOR_ROLE,
@@ -1313,929 +1304,6 @@ describe('Exchange.sol', function () {
       expect(await ERC20Contract.balanceOf(taker)).to.be.equal(123000000);
       expect(await ERC20Contract2.balanceOf(maker)).to.be.equal(456000000);
       expect(await ERC20Contract2.balanceOf(taker)).to.be.equal(0);
-    });
-  });
-
-  describe('matchOrders with royalties', function () {
-    it('should execute a complete match order between ERC721 and ERC20 tokens in primary market', async function () {
-      const {
-        deployer: maker, // making deployer the maker to sell in primary market
-      } = await loadFixture(deployFixtures);
-
-      await ERC721WithRoyaltyV2981.mint(maker.getAddress(), 1, [
-        await FeeRecipientsData(maker.getAddress(), 10000),
-      ]);
-
-      await ERC721WithRoyaltyV2981.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-      await ERC20Contract.mint(taker.getAddress(), 100000000000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        100000000000
-      );
-
-      expect(await ERC721WithRoyaltyV2981.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        100000000000
-      );
-      const makerAsset = await AssetERC721(ERC721WithRoyaltyV2981, 1);
-      const takerAsset = await AssetERC20(ERC20Contract, 100000000000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await ExchangeContractAsUser.matchOrders([
-        {
-          orderLeft,
-          signatureLeft: makerSig,
-          orderRight,
-          signatureRight: takerSig,
-        },
-      ]);
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(100000000000);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(1);
-      expect(await ERC721WithRoyaltyV2981.ownerOf(1)).to.be.equal(
-        await taker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(maker.getAddress())).to.be.equal(
-        98770000000
-      );
-
-      // check primary market protocol fee
-      expect(
-        await ERC20Contract.balanceOf(defaultFeeReceiver.getAddress())
-      ).to.be.equal(1230000000); // 123 * 100000000000 / 10000 = 1230000000
-    });
-
-    it('should not execute match order when royalties exceed 50%', async function () {
-      const {deployer: creator} = await loadFixture(deployFixtures);
-
-      // set royalty greater than 50%
-      await ERC721WithRoyaltyV2981.setRoyalties(1000000);
-
-      await ERC721WithRoyaltyV2981.mint(creator.getAddress(), 1, [
-        await FeeRecipientsData(creator.getAddress(), 10000),
-      ]);
-      await ERC721WithRoyaltyV2981.connect(creator).transferFrom(
-        creator.getAddress(),
-        maker.getAddress(),
-        1
-      );
-      await ERC721WithRoyaltyV2981.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-      await ERC20Contract.mint(taker.getAddress(), 1000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1000
-      );
-
-      expect(await ERC721WithRoyaltyV2981.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        1000
-      );
-      const makerAsset = await AssetERC721(ERC721WithRoyaltyV2981, 1);
-      const takerAsset = await AssetERC20(ERC20Contract, 1000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await expect(
-        ExchangeContractAsUser.matchOrders([
-          {
-            orderLeft,
-            signatureLeft: makerSig,
-            orderRight,
-            signatureRight: takerSig,
-          },
-        ])
-      ).to.be.revertedWith('Royalties are too high (>50%)');
-    });
-
-    it('should not execute match orders when royalties exceed 50% for token without IROYALTYUGC support', async function () {
-      // set royalty greater than 50%
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.setRoyalties(1000000);
-
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.mint(maker.getAddress(), 1, [
-        await FeeRecipientsData(maker.getAddress(), 10000),
-      ]);
-
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-      await ERC20Contract.mint(taker.getAddress(), 1000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1000
-      );
-
-      expect(await ERC721WithRoyaltyWithoutIROYALTYUGC.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        1000
-      );
-      const makerAsset = await AssetERC721(
-        ERC721WithRoyaltyWithoutIROYALTYUGC,
-        1
-      );
-      const takerAsset = await AssetERC20(ERC20Contract, 1000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await expect(
-        ExchangeContractAsUser.matchOrders([
-          {
-            orderLeft,
-            signatureLeft: makerSig,
-            orderRight,
-            signatureRight: takerSig,
-          },
-        ])
-      ).to.be.revertedWith('Royalties are too high (>50%)');
-    });
-
-    it('should not execute match orders when royalties exceed 50% for token without IROYALTYUGC and royalty.length != 1', async function () {
-      const {admin: receiver1, user: receiver2} = await loadFixture(
-        deployFixtures
-      );
-
-      // set royalty greater than 50%
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.setRoyalties(1000000);
-
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.mint(maker.getAddress(), 1, [
-        await FeeRecipientsData(receiver1.getAddress(), 3000),
-        await FeeRecipientsData(receiver2.getAddress(), 7000),
-      ]);
-
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-      await ERC20Contract.mint(taker.getAddress(), 1000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1000
-      );
-
-      expect(await ERC721WithRoyaltyWithoutIROYALTYUGC.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        1000
-      );
-      const makerAsset = await AssetERC721(
-        ERC721WithRoyaltyWithoutIROYALTYUGC,
-        1
-      );
-      const takerAsset = await AssetERC20(ERC20Contract, 1000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await expect(
-        ExchangeContractAsUser.matchOrders([
-          {
-            orderLeft,
-            signatureLeft: makerSig,
-            orderRight,
-            signatureRight: takerSig,
-          },
-        ])
-      ).to.be.revertedWith('Royalties are too high (>50%)');
-    });
-
-    it('should execute match orders for token without IROYALTYUGC support', async function () {
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.mint(maker.getAddress(), 1, [
-        await FeeRecipientsData(maker.getAddress(), 10000),
-      ]);
-
-      await ERC721WithRoyaltyWithoutIROYALTYUGC.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-      await ERC20Contract.mint(taker.getAddress(), 100000000000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        100000000000
-      );
-
-      expect(await ERC721WithRoyaltyWithoutIROYALTYUGC.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        100000000000
-      );
-      const makerAsset = await AssetERC721(
-        ERC721WithRoyaltyWithoutIROYALTYUGC,
-        1
-      );
-      const takerAsset = await AssetERC20(ERC20Contract, 100000000000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await ExchangeContractAsUser.matchOrders([
-        {
-          orderLeft,
-          signatureLeft: makerSig,
-          orderRight,
-          signatureRight: takerSig,
-        },
-      ]);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(100000000000);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(1);
-      expect(await ERC721WithRoyaltyWithoutIROYALTYUGC.ownerOf(1)).to.be.equal(
-        await taker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(maker.getAddress())).to.be.equal(
-        97500000000
-      );
-
-      // check protocol fee
-      expect(
-        await ERC20Contract.balanceOf(defaultFeeReceiver.getAddress())
-      ).to.be.equal(2500000000); // 250 * 10000000000 / 10000 = 2500000000
-    });
-
-    it('should execute a complete match order with external royalties provider(type 1)', async function () {
-      const {deployer: royaltyReceiver} = await loadFixture(deployFixtures);
-
-      await ERC721Contract.mint(maker.getAddress(), 1);
-
-      await ERC721Contract.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-
-      // set up royalties by token
-      await RoyaltiesRegistryAsDeployer.setRoyaltiesByToken(
-        await ERC721Contract.getAddress(),
-        [await LibPartData(royaltyReceiver, 2000)]
-      );
-
-      await ERC20Contract.mint(taker.getAddress(), 100000000000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        100000000000
-      );
-
-      expect(await ERC721Contract.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        100000000000
-      );
-      const makerAsset = await AssetERC721(ERC721Contract, 1);
-      const takerAsset = await AssetERC20(ERC20Contract, 100000000000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await ExchangeContractAsUser.matchOrders([
-        {
-          orderLeft,
-          signatureLeft: makerSig,
-          orderRight,
-          signatureRight: takerSig,
-        },
-      ]);
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(100000000000);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(1);
-      expect(await ERC721Contract.ownerOf(1)).to.be.equal(
-        await taker.getAddress()
-      );
-
-      // check protocol fee
-      expect(
-        await ERC20Contract.balanceOf(defaultFeeReceiver.getAddress())
-      ).to.be.equal(2500000000); // 250 * 10000000000 / 10000 = 2500000000
-
-      // check paid royalty
-      expect(
-        await ERC20Contract.balanceOf(royaltyReceiver.getAddress())
-      ).to.be.equal(20000000000); // 20% of the amount
-
-      expect(await ERC20Contract.balanceOf(maker.getAddress())).to.be.equal(
-        77500000000 // 100000000000 - royalty - protocolFee
-      );
-    });
-
-    it('should execute a complete match order with external royalties provider(type 2)', async function () {
-      const {deployer: royaltyReceiver} = await loadFixture(deployFixtures);
-
-      await ERC721Contract.mint(maker.getAddress(), 1);
-
-      await ERC721Contract.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-
-      // configuring royalties
-      await RoyaltiesProvider.initializeProvider(
-        await ERC721Contract.getAddress(),
-        1,
-        [await LibPartData(royaltyReceiver, 1000)]
-      );
-      await RoyaltiesRegistryAsDeployer.setProviderByToken(
-        await ERC721Contract.getAddress(),
-        RoyaltiesProvider.getAddress()
-      );
-
-      await ERC20Contract.mint(taker.getAddress(), 100000000000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        100000000000
-      );
-
-      expect(await ERC721Contract.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        100000000000
-      );
-      const makerAsset = await AssetERC721(ERC721Contract, 1);
-      const takerAsset = await AssetERC20(ERC20Contract, 100000000000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await ExchangeContractAsUser.matchOrders([
-        {
-          orderLeft,
-          signatureLeft: makerSig,
-          orderRight,
-          signatureRight: takerSig,
-        },
-      ]);
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(100000000000);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(1);
-      expect(await ERC721Contract.ownerOf(1)).to.be.equal(
-        await taker.getAddress()
-      );
-
-      // check protocol fee
-      expect(
-        await ERC20Contract.balanceOf(defaultFeeReceiver.getAddress())
-      ).to.be.equal(2500000000); // 250 * 10000000000 / 10000 = 2500000000
-
-      // check paid royalty
-      expect(
-        await ERC20Contract.balanceOf(royaltyReceiver.getAddress())
-      ).to.be.equal(10000000000); // 10% of the amount
-
-      expect(await ERC20Contract.balanceOf(maker.getAddress())).to.be.equal(
-        87500000000 // 100000000000 - royalty - protocolFee
-      );
-    });
-
-    it('should execute a complete match order with royalties 2981(type 3) transferred to royaltyReceiver', async function () {
-      const {
-        deployer: royaltyReceiver,
-
-        admin: receiver1,
-        user: receiver2,
-      } = await loadFixture(deployFixtures);
-
-      // set royalty
-      await ERC721WithRoyalty.setRoyalties(5000);
-
-      const fees = [
-        {account: receiver1.getAddress(), value: 4000},
-        {account: receiver2.getAddress(), value: 5000},
-      ];
-      await ERC721WithRoyalty.mint(maker.getAddress(), 1, fees);
-
-      await ERC721WithRoyalty.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-
-      // set up receiver
-      await ERC721WithRoyalty.setRoyaltiesReceiver(
-        1,
-        royaltyReceiver.getAddress()
-      );
-
-      await ERC20Contract.mint(taker.getAddress(), 100000000000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        100000000000
-      );
-
-      expect(await ERC721WithRoyalty.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        100000000000
-      );
-      const makerAsset = await AssetERC721(ERC721WithRoyalty, 1);
-      const takerAsset = await AssetERC20(ERC20Contract, 100000000000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await ExchangeContractAsUser.matchOrders([
-        {
-          orderLeft,
-          signatureLeft: makerSig,
-          orderRight,
-          signatureRight: takerSig,
-        },
-      ]);
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(100000000000);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(1);
-      expect(await ERC721WithRoyalty.ownerOf(1)).to.be.equal(
-        await taker.getAddress()
-      );
-
-      // check primary market protocol fee
-      expect(
-        await ERC20Contract.balanceOf(defaultFeeReceiver.getAddress())
-      ).to.be.equal(2500000000); // 250 * 100000000000 / 10000 = 2500000000
-
-      expect(await ERC20Contract.balanceOf(receiver1.getAddress())).to.be.equal(
-        0
-      );
-
-      expect(await ERC20Contract.balanceOf(receiver2.getAddress())).to.be.equal(
-        0
-      );
-
-      expect(
-        await ERC20Contract.balanceOf(royaltyReceiver.getAddress())
-      ).to.be.equal(
-        50000000000 // 5000 * 100000000000 / 10000 = 50000000000
-      );
-
-      expect(await ERC20Contract.balanceOf(maker.getAddress())).to.be.equal(
-        47500000000 // 100000000000 - royalty - protocolFee
-      );
-    });
-
-    it('should execute a complete match order with royalties 2981(type 3) transferred to fee recipients', async function () {
-      const {
-        deployer: royaltyReceiver,
-
-        admin: receiver1,
-        user: receiver2,
-      } = await loadFixture(deployFixtures);
-
-      await ERC721WithRoyaltyV2981.mint(maker.getAddress(), 1, [
-        await FeeRecipientsData(receiver1.getAddress(), 3000),
-        await FeeRecipientsData(receiver2.getAddress(), 7000),
-      ]);
-
-      await ERC721WithRoyaltyV2981.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-
-      // set up receiver
-      await ERC721WithRoyaltyV2981.setRoyaltiesReceiver(
-        1,
-        royaltyReceiver.getAddress()
-      );
-
-      await ERC20Contract.mint(taker.getAddress(), 100000000000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        100000000000
-      );
-
-      expect(await ERC721WithRoyaltyV2981.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        100000000000
-      );
-      const makerAsset = await AssetERC721(ERC721WithRoyaltyV2981, 1);
-      const takerAsset = await AssetERC20(ERC20Contract, 100000000000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await ExchangeContractAsUser.matchOrders([
-        {
-          orderLeft,
-          signatureLeft: makerSig,
-          orderRight,
-          signatureRight: takerSig,
-        },
-      ]);
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(100000000000);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(1);
-      expect(await ERC721WithRoyaltyV2981.ownerOf(1)).to.be.equal(
-        await taker.getAddress()
-      );
-
-      // check protocol fee
-      expect(
-        await ERC20Contract.balanceOf(defaultFeeReceiver.getAddress())
-      ).to.be.equal(2500000000); // 250 * 100000000000 / 10000 = 2500000000
-
-      expect(await ERC20Contract.balanceOf(receiver1.getAddress())).to.be.equal(
-        15000000000 // 1500 * 100000000000 / 10000 = 15000000000
-      );
-
-      expect(await ERC20Contract.balanceOf(receiver2.getAddress())).to.be.equal(
-        35000000000 // 3500 * 100000000000 / 10000 = 35000000000
-      );
-
-      expect(await ERC20Contract.balanceOf(maker.getAddress())).to.be.equal(
-        47500000000 // 100000000000 - royalty - protocolFee
-      );
-    });
-
-    it('should execute a complete match order without fee and royalties for privileged seller', async function () {
-      await ERC721WithRoyaltyV2981.mint(maker.getAddress(), 1, [
-        await FeeRecipientsData(maker.getAddress(), 10000),
-      ]);
-
-      await ERC721WithRoyaltyV2981.connect(maker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        1
-      );
-
-      // set up receiver
-      await ERC721WithRoyaltyV2981.setRoyaltiesReceiver(
-        1,
-        deployer.getAddress()
-      );
-
-      // grant exchange admin role to seller
-      await ExchangeContractAsDeployer.connect(admin).grantRole(
-        EXCHANGE_ADMIN_ROLE,
-        taker.getAddress()
-      );
-
-      await ERC20Contract.mint(taker.getAddress(), 100000000000);
-      await ERC20Contract.connect(taker).approve(
-        await ExchangeContractAsUser.getAddress(),
-        100000000000
-      );
-
-      expect(await ERC721WithRoyaltyV2981.ownerOf(1)).to.be.equal(
-        await maker.getAddress()
-      );
-      expect(await ERC20Contract.balanceOf(taker.getAddress())).to.be.equal(
-        100000000000
-      );
-      const makerAsset = await AssetERC721(ERC721WithRoyaltyV2981, 1);
-      const takerAsset = await AssetERC20(ERC20Contract, 100000000000);
-      const orderLeft = await OrderDefault(
-        maker,
-        makerAsset,
-        ZeroAddress,
-        takerAsset,
-        1,
-        0,
-        0
-      );
-      const orderRight = await OrderDefault(
-        taker,
-        takerAsset,
-        ZeroAddress,
-        makerAsset,
-        1,
-        0,
-        0
-      );
-      const makerSig = await signOrder(orderLeft, maker, OrderValidatorAsAdmin);
-      const takerSig = await signOrder(
-        orderRight,
-        taker,
-        OrderValidatorAsAdmin
-      );
-
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(0);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(0);
-
-      await ExchangeContractAsUser.matchOrders([
-        {
-          orderLeft,
-          signatureLeft: makerSig,
-          orderRight,
-          signatureRight: takerSig,
-        },
-      ]);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderLeft))
-      ).to.be.equal(100000000000);
-      expect(
-        await ExchangeContractAsUser.fills(hashKey(orderRight))
-      ).to.be.equal(1);
-      expect(await ERC721WithRoyaltyV2981.ownerOf(1)).to.be.equal(
-        await taker.getAddress()
-      );
-
-      // no protocol fee paid
-      expect(
-        await ERC20Contract.balanceOf(defaultFeeReceiver.getAddress())
-      ).to.be.equal(0);
-
-      // no royalties paid
-      expect(await ERC20Contract.balanceOf(deployer.getAddress())).to.be.equal(
-        0
-      );
-
-      expect(await ERC20Contract.balanceOf(maker.getAddress())).to.be.equal(
-        100000000000
-      );
     });
   });
